requirements.txt
pygame
Pillow
requests
flask
python-dotenv
numpy